requirements.txt
pygame
watchdog
Pillow
requests
scikit-learn
openai
piexif
numpy
sounddevice
wavio
soundfile
colorama